pyproject.toml
[project]
name = "warporacer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gymnasium>=1.2.3",
    "imageio[ffmpeg]>=2.37.3",
    "imgui-bundle>=1.92.700",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "pyglet>=2.1.14",
    "pyopengl>=3.1.10",
    "pyyaml>=6.0.3",
    "scikit-image>=0.26.0",
    "scipy>=1.17.1",
    "skrl>=2.0.0",
    "torch>=2.11.0",
    "typer>=0.24.1",
    "wandb>=0.26.0",
    "warp-lang>=1.12.1",
    "warp-nn>=0.1.0",
]
